==> jegil32/cluster_app_can2025/02-7_preparation_donnees_dv.ipynb <==
library(tidyverse)
library(factoextra)

# Charger les données
DataApp <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/01_app_2025.rds")
DataRCI <- readRDS("../../_SharedFolder_datagotchi_federal_2024/data/potGrowth/01_pilote_with_rci.rds")

# Charger le fichier des variables
source("01-1_variables.R")

df_dv <- DataApp %>%
  select(all_of(variables_vote))

df_RCI <- DataRCI %>%
  select(id, starts_with("rci"))


# Boucle pour afficher la table pour chaque variable
for (v in variables_vote) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_dv[[v]]))
    cat("\n\n")
  }
}

library(dplyr)

df_dv$dv_voteChoiceLPC <- NA
df_dv$dv_voteChoiceLPC[df_dv$dv_voteChoice == "lpc"] <- 1
df_dv$dv_voteChoiceLPC[df_dv$dv_voteChoice != "lpc"] <- 0

df_dv$dv_voteChoiceCPC <- NA
df_dv$dv_voteChoiceCPC[df_dv$dv_voteChoice == "cpc"] <- 1
df_dv$dv_voteChoiceCPC[df_dv$dv_voteChoice != "cpc"] <- 0

df_dv$dv_voteChoiceNDP <- NA
df_dv$dv_voteChoiceNDP[df_dv$dv_voteChoice == "ndp"] <- 1
df_dv$dv_voteChoiceNDP[df_dv$dv_voteChoice != "ndp"] <- 0

df_dv$dv_voteChoiceBQ <- NA
df_dv$dv_voteChoiceBQ[df_dv$dv_voteChoice == "bq"] <- 1
df_dv$dv_voteChoiceBQ[df_dv$dv_voteChoice != "bq"] <- 0

df_dv$dv_voteChoiceGPC <- NA
df_dv$dv_voteChoiceGPC[df_dv$dv_voteChoice == "gpc"] <- 1
df_dv$dv_voteChoiceGPC[df_dv$dv_voteChoice != "gpc"] <- 0

df_dv$dv_voteChoiceOther <- NA
df_dv$dv_voteChoiceOther[df_dv$dv_voteChoice == "other"] <- 1
df_dv$dv_voteChoiceOther[df_dv$dv_voteChoice != "other"] <- 0

#Création d'une variable de solidité du vote
library(dplyr)

df_RCI <- df_RCI %>%
  rowwise() %>%
  mutate(dv_solidity = max(c_across(starts_with("rci_")), na.rm = TRUE)) %>%
  ungroup()


#création d'une variable de turnout binaire
library(dplyr)

df_dv$dv_turnout_bin <- NA
df_dv$dv_turnout_bin[df_dv$dv_turnout >= 0.9] <- 1
df_dv$dv_turnout_bin[df_dv$dv_turnout < 0.9] <- 0
table(df_dv$dv_turnout_bin)

table(df_RCI$dv_solidity)
table(df_RCI$rci_LPC)

df_dv <- left_join(df_dv, df_RCI, by = "id")


# Sauvegarder les données préparées
saveRDS(df_dv, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_app_2025_dv.rds")
